requirements.txt
langchain>=0.2.0
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
python-dotenv
tavily-python
sentence-transformers
